+++ ServerDevcommands/ServerExecution.cs
@@ -1,5 +1,7 @@
 using System.Collections.Generic;
+using System.Linq;
 using HarmonyLib;
+
 namespace ServerDevcommands;
 
 [HarmonyPatch(typeof(Terminal), nameof(Terminal.AddString), new[] { typeof(string) })]
@@ -38,7 +40,7 @@ public class ServerExecution
   {
     var server = ZNet.instance.GetServerRPC();
     Console.instance.AddString("Sending command: " + command);
-    if (server != null) server.Invoke(RPC_Command, new[] { command });
+    if (server != null) server.Invoke(RPC_Command, command);
   }
   ///<summary>Sends command to the server so that it can be executed there.</summary>
   public static void Send(IEnumerable<string> args) => Send(string.Join(" ", args));
@@ -46,6 +48,7 @@ public class ServerExecution
   public static void Send(Terminal.ConsoleEventArgs args) => Send(args.Args);
 
   public static string RPC_Command = "DEV_Command";
+  public static string RPC_Pins = "DEV_Pins";
   private static bool IsAllowed(ZRpc rpc, string command)
   {
     var zNet = ZNet.instance;
@@ -72,11 +75,35 @@ public class ServerExecution
       Console.instance.TryRunCommand(command);
     RedirectOutput.Target = null;
   }
+  private static void RPC_Do_Pins(ZRpc rpc, string data)
+  {
+    var pins = Parse.Split(data, '|').Select(Parse.VectorXZY).ToArray();
+    var findPins = Console.instance.m_findPins;
+    foreach (var pin in findPins)
+      Minimap.instance?.RemovePin(pin);
+    findPins.Clear();
+    if (pins.Length == 1)
+    {
+      Chat.instance?.SendPing(pins[0]);
+      return;
+    }
+    foreach (var pos in pins)
+    {
+      var pin = Minimap.instance?.AddPin(pos, Minimap.PinType.Icon3, "", false, false, Player.m_localPlayer.GetPlayerID());
+      if (pin != null)
+        findPins.Add(pin);
+    }
+  }
+
   static void Postfix(ZNet __instance, ZRpc rpc)
   {
     if (__instance.IsServer())
     {
       rpc.Register<string>(RPC_Command, new(RPC_Do_Command));
+    }
+    else
+    {
+      rpc.Register<string>(RPC_Pins, new(RPC_Do_Pins));
     }
   }
 }