+++ publish/CHANGELOG.md
@@ -1,5 +1,6 @@
 - v1.46
   - Adds a new command `players` to print online players.
+  - Changes the `find` command to server side.
 
 - v1.45
   - Adds server side command logging when installed on both client and server.